Treasure Hunt in C
[Treasure Hunt ] Write a program to explore the below array for a treasure. The
values in the array are clues. Each cell contains an integer between 11 and 55; for
each value the ten's digit represents the row number and the unit's digit
represents the column number of the cell containing the next clue. Starting in
the upper left corner (at 1,1), use the clues to guide your search of the array. (The
first three clues are 11, 34, 42). The treasure is a cell whose value is the same as
its coordinates. Your program must first read the treasure map data into a 5 by 5
array. Your program should output the cells it visits during its search, and a
message indicating where you found the treasure.